backend/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
cors
sqlalchemy
asyncpg
greenlet
